refactor: useDiaryRegister 훅 생성
